// Sipro/Sipro/ClientApp/src/app/components/adquisicion/model/Adquisicion.ts
export class Adquisicion{
    id: number;
    tipoNombre: string;
    tipoAdquisicion: number;
    categoriaNombre: string;
    categoriaAdquisicion: number;
    unidadMedida: string;
    cantidad: number;
    precioUnitario: number;
    total: number;
    nog: number;
    numeroContrato: string;
    montoContrato: number;
    preparacionDocPlanificado: string;
    preparacionDocReal: string;
    lanzamientoEventoPlanificado: string;
    lanzamientoEventoReal: string;
    recepcionOfertasPlanificado: string;
    recepcionOfertasReal: string;
    adjudicacionPlanificado: string;
    adjudicacionReal: string;
    firmaContratoPlanificado: string;
    firmaContratoReal: string;
    pagos: string;
    tipoRevision: number;
    tipoRevisionNombre: string;
    objetoId: number;
    objetoTipo: number;
}